fix UnrootedSmallParsimony and NearestNeighborsOfTree (ignore line order)
ensure that the order of the lines in the input does not matter (i.e. do
not check line number to determine whether we are adding a new leaf node
and its parent or a pair of internal nodes). The original code relied
heavily in the order of the inputs.