Fix temporary root node insertion for unrooted trees
Before, it was assumed that the last two nodes are connected so they
were used as insertion point for the temporary root. However, it is
possible that the last 2 nodes are not connected. For the new insertion
point selection strategy, the last added node and one of its connected
nodes will be chosen as the insertion point.