Transfer initialization logic of character trees to separate code block
When the first line of adjacency list contains both integers, the
previous code fails to create the character trees. To prevent this, the
new code will first go through the lines until it finds a dna string;
then, it computes the length from this. After it retrieves the length
will it only begin really processing the lines.