Add consideration for proteins with duplicate masses (i.e. I/L, K/Q)
The amino acid string that the previous code returns does not know
whether to select I or L for mass 113 / K or Q for mass 128. To solve
this, we take note of the starting index (in the proteome) of the best
peptide and then extract the actual string from the proteome.